+++ src/features/chat/ui/Chat/Chat.module.css
@@ -5,7 +5,7 @@
   height: 100%;
   padding: 0 2rem;
 
-  & .pageInner {
+  .pageInner {
     display: flex;
     flex-direction: column;
     width: 100%;
