--- src/features/chat/ui/ChatBottomBar/index.ts
@@ -1,1 +1,0 @@
-export * from "./ChatBottomBar";
