+++ src/features/chat/ui/ChatHeader/ChatHeader.module.css
@@ -5,12 +5,12 @@
   padding: 0.8rem 0.6rem;
   border-bottom: 1px solid var(--mantine-color-dark-4);
 
-  & .leftSection {
+  .leftSection {
     display: flex;
     align-items: center;
     gap: 0.2rem;
 
-    & .siteTitle {
+    .siteTitle {
       font-size: 1.5rem;
       line-height: 1.5rem;
       font-weight: 700;
@@ -20,7 +20,7 @@
       border-radius: var(--mantine-radius-md);
     }
 
-    & .separator {
+    .separator {
       font-size: 1.5rem;
       line-height: 0;
       font-weight: 700;
@@ -29,13 +29,13 @@
     }
   }
 
-  & .clearButton {
+  .clearButton {
     width: 2rem;
     height: 2rem;
     padding: 0.25rem;
     border-radius: var(--mantine-radius-md);
 
-    & * {
+    * {
       color: var(--mantine-color-dark-2);
     }
