+++ src/features/chat/ui/ChatMessages/ChatMessages.module.css
@@ -7,7 +7,7 @@
   border-bottom: 1px solid var(--mantine-color-dark-4);
   overflow-y: scroll;
 
-  & .messagesContainer {
+  .messagesContainer {
     display: flex;
     flex-direction: column;
     justify-content: flex-end;
