+++ src/features/chat/ui/ChatModelPicker/ChatModelPicker.module.css
@@ -8,7 +8,7 @@
     background-color: transparent;
   }
 
-  & .modelName {
+  .modelName {
     font-size: 1.5rem;
     line-height: 1.5rem;
     font-weight: 500;
@@ -17,7 +17,7 @@
     margin: 0;
   }
 
-  & svg {
+  svg {
     position: relative;
     top: 0.2rem;
     line-height: 0;
