+++ src/features/chat/ui/ChatModelPicker/index.ts
@@ -1,0 +1,1 @@
+export * from "./ChatModelPicker";
