--- src/features/chat/ui/ChatModelSpotlight/index.ts
@@ -1,1 +1,0 @@
-export * from "./ChatModelSpotlight";
