+++ src/features/chat/ui/ChatPrompt/ChatPrompt.module.css
@@ -3,19 +3,19 @@
   gap: 0.5rem;
   padding: 0.6rem 0;
 
-  & .inputRoot {
+  .inputRoot {
     flex-grow: 1;
 
-    & .input {
+    .input {
       height: 2.5rem;
       padding: 0 1.2rem;
       border-radius: 10rem;
     }
 
-    & .inputRightSection {
+    .inputRightSection {
       width: 2.5rem;
 
-      & .clearButton {
+      .clearButton {
         width: 2rem;
         height: 2rem;
         padding: 0.15rem;
@@ -24,12 +24,12 @@
     }
   }
 
-  & .submitButtonRoot {
+  .submitButtonRoot {
     width: 3.5rem;
     height: 2.5rem;
     border-radius: 10rem;
 
-    & .submitIcon {
+    .submitIcon {
       width: 1.2rem;
       height: 1.2rem;
     }
