+++ src/features/chat/ui/ChatPrompt/index.ts
@@ -1,0 +1,1 @@
+export * from "./ChatPrompt";
