--- src/features/chat/ui/ChatPullModelProgress/index.ts
@@ -1,1 +1,0 @@
-export * from "./ChatPullModelProgress";
