+++ src/features/chat/ui/ChatPullProgress/ChatPullProgress.module.css
@@ -12,7 +12,7 @@
     top: -0.2rem;
     color: var(--mantine-color-dark-2);
 
-    & p {
+    p {
       display: flex;
       gap: 0.3rem;
       align-items: center;
